Menu + DialogManager Update
+ Menu scene and logic added
- BUG - The text does not go where directed, instead going to one
particular location. @Bio-Roxas will be fixing this.
+ DialogManager.cs - Updated the manager to account for populating the
dialog box with text asset text. Randomly chooses text dialog from the
dialogs the character can have with someone. If none are available, we
will show three dots.
+ DialogHolder.cs - This holds the dialogs to choose from as well as
marks if the text is important or not. If it is, then we can not close
the dialog box early.